+++ PEPATACr/tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(PEPATACr)
+test_check("PEPATACr")